fix: 修复 14+ 个 Kotlin 编译错误
1. MainActivity: StateFlow<Boolean> → 先 collectAsState 再使用
2. MemorialDay: return 不能在表达式体中使用 → 改为块体
3. ReminderSettings: from targetDate → targetDate (参数名语法)
4. ImageService: fromHex 冲突 + LinearGradient 歧义 → 私有函数改名
5. LunarCalendarService: 重复 var offset → 移除重复声明
6. DateCalculatorScreen: HorizontalDivider → Divider (M3 1.1 API)
7. DateCalculatorScreen: QuickConvertButton 参数顺序调换
8. CardEditorScreen: Dp.toInt() 不存在 → 直接用 Dp
9. HomeScreen: Modifier.align 缺少 BoxScope → 改用 Box 包裹
10. MemorialCardView: Dp.toInt() 不存在 → 直接用 Dp
11. TemplatePickerScreen: items 变量名与 LazyListScope 冲突 → 改名
12. StorageService: settings 私有 setter → 改为公开

+++ app/src/main/java/com/memorialday/app/ui/detail/CardEditorScreen.kt
@@ -274,13 +274,13 @@ private fun PreviewCard(
             modifier = Modifier
                 .fillMaxWidth()
                 .height(200.dp),
-            shape = RoundedCornerShape(cornerRadius.dp.toInt()),
-            shadowElevation = shadowRadius.dp.toInt().coerceAtMost(8).dp,
+            shape = RoundedCornerShape(cornerRadius.dp),
+            shadowElevation = shadowRadius.dp.coerceAtMost(8.dp),
             color = Color.Transparent
         ) {
             Box(
                 modifier = Modifier
-                    .background(brush = bg, shape = RoundedCornerShape(cornerRadius.dp.toInt())),
+                    .background(brush = bg, shape = RoundedCornerShape(cornerRadius.dp)),
                 contentAlignment = Alignment.Center
             ) {
                 Column(horizontalAlignment = Alignment.CenterHorizontally) {
